--- src/components/SectionA/SectionA.css
@@ -1,61 +1,0 @@
-/** @format */
-.section-a {
-  width: 100%;
-  max-width: 100%;
-}
-
-@media (min-width: 991px) {
-  .grid-container-a {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: 100%;
-    grid-gap: 30px;
-    padding: 30px 0;
-  }
-}
-@media (max-width: 992px) {
-  .grid-item-a-text-box {
-    margin-bottom: 30px;
-  }
-  .grid-item-a-image-box {
-    margin-bottom: 30px;
-  }
-}
-.grid-item-a {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  transition: all 0.4s;
-  box-sizing: border-box;
-  border: 2px solid transparent;
-}
-
-#grid-item-overlay {
-  height: 100%;
-  width: 100%;
-}
-
-.grid-item-a:hover,
-.grid-item-a:active {
-  background: transparent;
-  color: var(--dark);
-  cursor: pointer;
-}
-
-.grid-item-a-text-box {
-  padding: 20px;
-  text-align: left;
-  background: var(--dark);
-  color: white;
-}
-.grid-item-a-text-box:hover, .grid-item-a-text-box:focus, .grid-item-a-text-box:active {
-  border: 2px dashed grey;
-}
-
-.grid-item-a-img {
-  width: 100%;
-  height: 100%;
-}
-
-/*# sourceMappingURL=SectionA.css.map */
